requirements.txt
google-cloud-bigquery
db-dtypes
pygwalker
pandas
streamlit
